requirements.txt
kafka-python
requests
elasticsearch==7.17.9